rentpoints/requirement.txt
psycopg2==2.7.3
django==2.1.1
djangorestframework==3.8.2